Fixed run-on note glitch
Seems to be a timing issue where the gate would be changed by a bus write during the synth processing, causing an impossible state. The fix was to add a synchronization buffer so that the gate is synchronous to the Synth Clock instead of the bus clock.
Also improved some things in the ADSR